Update Float type to f64 for improved precision and add Float64 capability in build script

+++ src/lib.rs
@@ -2,7 +2,7 @@
 #![allow(unexpected_cfgs)]
 
 pub mod kernels;
-pub type Float = f32;
+pub type Float = f64;
 
 #[cfg(not(target_arch = "spirv"))]
 mod shader_load;
